[#8] - (refactor) 디렉토리 구조 개선
- 도메인 기반 디렉토리 구조 변경
- 레거시 API 비활성화
- 레거시 API description 추가
