+++ app/api/stock/stock_service.py
